+++ test/Serilog.Sinks.HttpTests/DurableHttpSinkGivenCodeConfigurationShould.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Microsoft.Extensions.Configuration;
 using Serilog.Core;
 using Serilog.Sinks.Http.BatchFormatters;
 using Serilog.Sinks.Http.TextFormatters;
@@ -11,6 +12,8 @@ namespace Serilog
         public DurableHttpSinkGivenCodeConfigurationShould()
         {
             DeleteBufferFiles();
+            
+            var configuration = new ConfigurationBuilder().Build();
 
 #pragma warning disable CS0618 // Type or member is obsolete
 
@@ -23,12 +26,17 @@ namespace Serilog
                     period: TimeSpan.FromMilliseconds(1),
                     textFormatter: new NormalRenderedTextFormatter(),
                     batchFormatter: new DefaultBatchFormatter(),
-                    httpClient: new HttpClientMock())
+                    httpClient: new HttpClientMock(),
+                    configuration: configuration)
                 .CreateLogger();
 
 #pragma warning restore CS0618 // Type or member is obsolete
+
+            Configuration = configuration;
         }
 
         protected override Logger Logger { get; }
+
+        protected override IConfiguration Configuration { get; }
     }
 }